backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

pydantic==2.10.3
pydantic-settings==2.6.1

python-multipart==0.0.12

sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet>=3.0.0

faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
numpy==1.26.4

google-generativeai==0.8.3

orjson==3.10.12
python-dateutil==2.9.0.post0

httpx==0.28.0
tenacity==9.0.0

pytest==8.3.4
pytest-asyncio==0.24.0
